+++ whatslab2/src/componentes/ComponenteMensagem/ComponenteMensagem.js
@@ -1,0 +1,24 @@
+import React from "react";
+
+export default class ComponenteMensagem extends React.Component {
+  state = {
+    //---------------------------------------------------------
+    // Array de usuario e mensagens
+    mensagens: [],
+    //---------------------------------------------------------
+    //Valores de estado dos inputs
+    valorInputUsuario: "",
+    valorInputMensagem: ""
+  };
+  
+  render() {
+    const componentesMensagens = this.state.mensagens.map((mensagem) => {
+      return (
+        <p>
+          {mensagem.usuario}: {mensagem.textoMensagem}
+        </p>
+      );
+    });
+    return <div className="App">{componentesMensagens}</div>;
+  }
+}
